requirements.txt
Pillow>=9.0.0
Flask>=2.0
numpy>=1.22
sounddevice>=0.4.5
